chore: move game-session test helpers to top-level client utils
