--- client/src/pages/GameSession/utils/testing/index.ts
@@ -1,2 +1,0 @@
-export * from './domElementGetters';
-export * from './expectHelpers';
