+++ client/src/utils/testing/dom-element-getters/index.ts
@@ -1,0 +1,1 @@
+export * from './game-session';
