+++ client/src/utils/testing/expect-helpers/common.ts
@@ -1,0 +1,18 @@
+import { type Screen } from '@testing-library/react';
+
+export async function expectHeadingToBeVisible({
+  screenRef,
+  level = 2,
+  name = /Connect Four/,
+}: {
+  screenRef: Screen;
+  level?: number;
+  name?: RegExp | string;
+}) {
+  const headingEl = await screenRef.findByRole('heading', {
+    level,
+    name,
+  });
+
+  expect(headingEl).toBeVisible();
+}
