+++ client/src/utils/testing/index.ts
@@ -1,2 +1,4 @@
+export * from './dom-element-getters';
+export * from './expect-helpers';
 export * from './fetchMocks';
 export * from './WithMemoryRouter';
